week02/requirements.txt
python-dotenv
requests
lxml